backend/requirements.txt
Django==5.1
psycopg[binary]
djangorestframework==3.15.2
django-cors-headers
Pillow
channels
channels_redis
daphne